Move method to where it's used
